Add RIT support - Ugh
So RIT is a perfect example of race conditions and positive feedback
We can *only* adjust it by an offset up or down, when we send the
message, it takes time to process, and we can't clear it unless
RIT or XIT is enabled (despite being able to set it).

What I've done is implement a simplified Nagle's algorithm.
If there's no RIT adjustment pending, a new one is enqueued.
If one is enqueued but not sent, it is simply updated.
If one is enqueued *and* sent, it is stored in a buffer.
If the buffer already exists, it is simply updated
Once we get the IF confirming a change, we place the buffered
version into the queue.

This is basically the same thing as "inhibit the sending of new TCP
segments when new outgoing data arrives from the user if any previously
transmitted data on the connection remains unacknowledged."

Just replace "TCP segments" with "RIT adjustments" and done.