More overhaul
The cache fill is now as complete as I can get it, but it's getting
overly complex again.  I think I need to put the correlation data
into the code somewhere and let it sort itself out rather than me
having to fumble my way though getting it.

This introduces the current* values which are intended to be used
by things like loggers and wsjt-x that only care about what it's
receiving and where and what happens if it toggles PTT.  The idea
is that rigctld emultation will exclusively use current* values.